# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Nifty100 Data Foundation API",
    "description": "Internal API over the Nifty100 fundamentals/screener/analytics data foundation.",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/companies": {
      "get": {
        "tags": [
          "companies"
        ],
        "summary": "List Companies",
        "description": "List companies.",
        "operationId": "list_companies_api_v1_companies_get",
        "parameters": [
          {
            "name": "sector",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by broad_sector",
              "title": "Sector"
            },
            "description": "Filter by broad_sector"
          },
          {
            "name": "market_cap_category",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by market_cap_category",
              "title": "Market Cap Category"
            },
            "description": "Filter by market_cap_category"
          },
          {
            "name": "search",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Partial match on company name or ticker",
              "title": "Search"
            },
            "description": "Partial match on company name or ticker"
          },
          {
            "name": "db_path",
            "in": "query",
            "required": false,
            "schema": {
              "default": "/home/claude/work/extracted/Nifty100_Data_Foundation/data/nifty100.db",
              "title": "Db Path"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/companies/{ticker}": {
      "get": {
        "tags": [
          "companies"
        ],
        "summary": "Get Company",
        "description": "Get company.",
        "operationId": "get_company_api_v1_companies__ticker__get",
        "parameters": [
          {
            "name": "ticker",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Ticker"
            }
          },
          {
            "name": "db_path",
            "in": "query",
            "required": false,
            "schema": {
              "default": "/home/claude/work/extracted/Nifty100_Data_Foundation/data/nifty100.db",
              "title": "Db Path"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/companies/{ticker}/pl": {
      "get": {
        "tags": [
          "companies"
        ],
        "summary": "Company Pl",
        "description": "Company pl.",
        "operationId": "company_pl_api_v1_companies__ticker__pl_get",
        "parameters": [
          {
            "name": "ticker",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Ticker"
            }
          },
          {
            "name": "from_year",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "YYYY or YYYY-MM",
              "title": "From Year"
            },
            "description": "YYYY or YYYY-MM"
          },
          {
            "name": "to_year",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "YYYY or YYYY-MM",
              "title": "To Year"
            },
            "description": "YYYY or YYYY-MM"
          },
          {
            "name": "db_path",
            "in": "query",
            "required": false,
            "schema": {
              "default": "/home/claude/work/extracted/Nifty100_Data_Foundation/data/nifty100.db",
              "title": "Db Path"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/companies/{ticker}/bs": {
      "get": {
        "tags": [
          "companies"
        ],
        "summary": "Company Balance Sheet",
        "description": "Company balance sheet.",
        "operationId": "company_balance_sheet_api_v1_companies__ticker__bs_get",
        "parameters": [
          {
            "name": "ticker",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Ticker"
            }
          },
          {
            "name": "from_year",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "YYYY or YYYY-MM",
              "title": "From Year"
            },
            "description": "YYYY or YYYY-MM"
          },
          {
            "name": "to_year",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "YYYY or YYYY-MM",
              "title": "To Year"
            },
            "description": "YYYY or YYYY-MM"
          },
          {
            "name": "db_path",
            "in": "query",
            "required": false,
            "schema": {
              "default": "/home/claude/work/extracted/Nifty100_Data_Foundation/data/nifty100.db",
              "title": "Db Path"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/companies/{ticker}/cashflow": {
      "get": {
        "tags": [
          "companies"
        ],
        "summary": "Company Cashflow",
        "description": "Company cashflow.",
        "operationId": "company_cashflow_api_v1_companies__ticker__cashflow_get",
        "parameters": [
          {
            "name": "ticker",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Ticker"
            }
          },
          {
            "name": "from_year",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "YYYY or YYYY-MM",
              "title": "From Year"
            },
            "description": "YYYY or YYYY-MM"
          },
          {
            "name": "to_year",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "YYYY or YYYY-MM",
              "title": "To Year"
            },
            "description": "YYYY or YYYY-MM"
          },
          {
            "name": "db_path",
            "in": "query",
            "required": false,
            "schema": {
              "default": "/home/claude/work/extracted/Nifty100_Data_Foundation/data/nifty100.db",
              "title": "Db Path"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/companies/{ticker}/ratios": {
      "get": {
        "tags": [
          "companies"
        ],
        "summary": "Company Ratios",
        "description": "Company ratios.",
        "operationId": "company_ratios_api_v1_companies__ticker__ratios_get",
        "parameters": [
          {
            "name": "ticker",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Ticker"
            }
          },
          {
            "name": "year",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Return a single year's ratios (YYYY or YYYY-MM)",
              "title": "Year"
            },
            "description": "Return a single year's ratios (YYYY or YYYY-MM)"
          },
          {
            "name": "db_path",
            "in": "query",
            "required": false,
            "schema": {
              "default": "/home/claude/work/extracted/Nifty100_Data_Foundation/data/nifty100.db",
              "title": "Db Path"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/companies/{ticker}/tearsheet": {
      "get": {
        "tags": [
          "companies"
        ],
        "summary": "Company Tearsheet",
        "description": "Company tearsheet.",
        "operationId": "company_tearsheet_api_v1_companies__ticker__tearsheet_get",
        "parameters": [
          {
            "name": "ticker",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Ticker"
            }
          },
          {
            "name": "db_path",
            "in": "query",
            "required": false,
            "schema": {
              "default": "/home/claude/work/extracted/Nifty100_Data_Foundation/data/nifty100.db",
              "title": "Db Path"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/companies/{ticker}/peers/compare": {
      "get": {
        "tags": [
          "companies"
        ],
        "summary": "Company Peer Radar",
        "description": "Company peer radar.",
        "operationId": "company_peer_radar_api_v1_companies__ticker__peers_compare_get",
        "parameters": [
          {
            "name": "ticker",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Ticker"
            }
          },
          {
            "name": "db_path",
            "in": "query",
            "required": false,
            "schema": {
              "default": "/home/claude/work/extracted/Nifty100_Data_Foundation/data/nifty100.db",
              "title": "Db Path"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/companies/{ticker}/documents": {
      "get": {
        "tags": [
          "companies"
        ],
        "summary": "Company Documents",
        "description": "Company documents.",
        "operationId": "company_documents_api_v1_companies__ticker__documents_get",
        "parameters": [
          {
            "name": "ticker",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Ticker"
            }
          },
          {
            "name": "db_path",
            "in": "query",
            "required": false,
            "schema": {
              "default": "/home/claude/work/extracted/Nifty100_Data_Foundation/data/nifty100.db",
              "title": "Db Path"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/sectors/{broad_sector}/companies": {
      "get": {
        "tags": [
          "sectors"
        ],
        "summary": "Companies In Sector",
        "description": "All companies in a sector, with latest-year KPIs. 404 for an unknown sector.",
        "operationId": "companies_in_sector_api_v1_sectors__broad_sector__companies_get",
        "parameters": [
          {
            "name": "broad_sector",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Broad Sector"
            }
          },
          {
            "name": "db_path",
            "in": "query",
            "required": false,
            "schema": {
              "default": "/home/claude/work/extracted/Nifty100_Data_Foundation/data/nifty100.db",
              "title": "Db Path"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}